requirements.txt
requests>=2.31.0
PyYAML>=6.0
openai>=1.0.0
anthropic>=0.3.0
python-dotenv>=1.0.0
atproto>=0.0.34